[#277] fixed conditions to prevent blink in compact event chips
